backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
PyMySQL
PyJWT
python-dotenv
Werkzeug
requests
uvicorn

